feat(nextcloud): epubviewerアップデート

+++ nextcloud/custom_apps/epubviewer/lib/Controller/PreferenceController.php
@@ -4,89 +4,88 @@ namespace OCA\Epubviewer\Controller;
 
 use OCA\Epubviewer\Service\PreferenceService;
 use OCP\AppFramework\Controller;
+use OCP\AppFramework\Http\Attribute\NoAdminRequired;
+use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
 use OCP\AppFramework\Http\JSONResponse;
 use OCP\IRequest;
 
 class PreferenceController extends Controller {
 
-	private $preferenceService;
-
 	/**
 	 * @param string $appName
 	 * @param IRequest $request
-	 * @param PreferenceService $preferenceService
+	 * @param ?PreferenceService $preferenceService
 	 */
 	public function __construct($appName,
 		IRequest $request,
-		PreferenceService $preferenceService) {
+		private ?PreferenceService $preferenceService,
+		private ?string $userId) {
 
 		parent::__construct($appName, $request);
-		$this->preferenceService = $preferenceService;
 	}
 
 	/**
-	 * @brief return preference for $fileId
-	 *
-	 * @NoAdminRequired
-	 * @NoCSRFRequired
-	 *
 	 * @param string $scope
 	 * @param int $fileId
 	 * @param string $name if null, return all preferences for $scope + $fileId
-	 *
-	 * @return array|JSONResponse
 	 */
-	public function get($scope, $fileId, $name) {
+	#[NoAdminRequired]
+	#[NoCSRFRequired]
+	public function get(string $scope, int $fileId, string $name): array|JSONResponse {
+		if ($this->userId === null) {
+			return new JSONResponse([]);
+		}
 		return $this->preferenceService->get($scope, $fileId, $name);
 	}
 
 	/**
 	 * @brief write preference for $fileId
 	 *
-	 * @NoAdminRequired
-	 * @NoCSRFRequired
-	 *
 	 * @param string $scope
 	 * @param int $fileId
 	 * @param string $name
 	 * @param string $value
-	 *
-	 * @return array|JSONResponse
 	 */
-	public function set($scope, $fileId, $name, $value) {
-		return $this->preferenceService->set($scope, $fileId, $name, $value);
+	#[NoAdminRequired]
+	#[NoCSRFRequired]
+	public function set(string $scope, int $fileId, string $name, string $value): void {
+		if ($this->userId === null) {
+			return;
+		}
+		$this->preferenceService->set($scope, $fileId, $name, $value);
 	}
 
 
 	/**
 	 * @brief return default preference
 	 *
-	 * @NoAdminRequired
-	 * @NoCSRFRequired
-	 *
 	 * @param string $scope
 	 * @param string $name if null, return all default preferences for scope
-	 *
-	 * @return array|JSONResponse
 	 */
-	public function getDefault($scope, $name) {
+	#[NoAdminRequired]
+	#[NoCSRFRequired]
+	public function getDefault(string $scope, string $name): array|JSONResponse {
+		if ($this->userId === null) {
+			return new JSONResponse([]);
+		}
 		return $this->preferenceService->getDefault($scope, $name);
 	}
 
 	/**
 	 * @brief write default preference
 	 *
-	 * @NoAdminRequired
-	 * @NoCSRFRequired
-	 *
 	 * @param string $scope
 	 * @param string $name
 	 * @param string $value
-	 *
-	 * @return array|JSONResponse
 	 */
-	public function setDefault($scope, $name, $value) {
-		return $this->preferenceService->setDefault($scope, $name, $value);
+	#[NoAdminRequired]
+	#[NoCSRFRequired]
+	public function setDefault(string $scope, string $name, string $value): void {
+		if ($this->userId === null) {
+			return;
+		}
+
+		$this->preferenceService->setDefault($scope, $name, $value);
 	}
 
 	/**
@@ -96,7 +95,10 @@ class PreferenceController extends Controller {
 	 * @param int $fileId
 	 * @param string $name
 	 */
-	public function delete($scope, $fileId, $name): void {
+	public function delete(string $scope, int $fileId, string $name): void {
+		if ($this->userId === null) {
+			return;
+		}
 		$this->preferenceService->delete($scope, $fileId, $name);
 	}
 
@@ -106,7 +108,10 @@ class PreferenceController extends Controller {
 	 * @param $scope
 	 * @param $name
 	 */
-	public function deleteDefault($scope, $name): void {
+	public function deleteDefault(string $scope, string $name): void {
+		if ($this->userId === null) {
+			return;
+		}
 		$this->preferenceService->deleteDefault($scope, $name);
 	}
 }
